Add quotes to all parameter accesses Add comment to setCertValidity Move PKCS#12 bundle creation from client signing into own function (now called after signing) Fix start date argument Swap some arguments (subject first, then required CA)
git-svn-id: http://svn.hazelnut4.nuts/openssl-ca@18 3c313b17-921d-4e24-803b-4865af1d1567

+++ ca_lib.sh
@@ -60,8 +60,17 @@ function cahelp()
     echo
 }
 
+
+# Set the validity dates of the next certificate to be signed. The the dates
+# are inclusive, meaning that the the validity starts at 0:00:00 and ends
+# at 23:59:59.
+#
+# This function is automatically called before signing a client or CA
+# certificate. It is not called for a self-signed root-certificate, for
+# which the validity must be set in this script.
 function setCertValidity()
 {
+    local firstLoop=1
     local inputStartDate=""
     local inputEndDate=""
 
@@ -69,7 +78,8 @@ function setCertValidity()
     local inputStartMM=
     local inputStartDD=
 
-    while [[ ! "${START_DATE}" =~ ^([0-9][0-9])(0[0-9]|1[012])([012][0-9]|3[01])$ ]]; do
+    while [[ ! "${START_DATE}" =~ ^([0-9][0-9])(0[0-9]|1[012])([012][0-9]|3[01])$ || $firstLoop -eq 1 ]]; do
+        firstLoop=0
         read -p "Not valid before (YYMMDD) [${START_DATE}]: " inputStartDate
         START_DATE=${inputStartDate:-$START_DATE}
     done
@@ -78,7 +88,9 @@ function setCertValidity()
     inputStartMM=${BASH_REMATCH[2]}
     inputStartDD=${BASH_REMATCH[3]}
 
-    while [[ ! "${inputEndDate}" =~ ^([0-9][0-9])(0[0-9]|1[012])([012][0-9]|3[01])$ || $END_DATE -lt $START_DATE ]]; do
+    firstLoop=1
+    while [[ ! "${END_DATE}" =~ ^([0-9][0-9])(0[0-9]|1[012])([012][0-9]|3[01])$ || $END_DATE -lt $START_DATE || $firstLoop -eq 1 ]]; do
+        firstLoop=0
         END_DATE=$((inputStartYY+DEFAULT_EXPIRY_YEARS))
         END_DATE="$END_DATE$inputStartMM$inputStartDD"
 
@@ -86,7 +98,7 @@ function setCertValidity()
         END_DATE=${inputEndDate:-$END_DATE}
     done
 }
-setCertValidity
+
 
 # Create a new CA key and create the necessary database structure. If the
 # CA key already exists, the process will be aborted.
@@ -103,7 +115,7 @@ setCertValidity
 #
 function createCA()
 {
-    local caName=$1
+    local caName="$1"
     local caConf="$CNF_ROOT/ca.$caName.conf"
     local outFile="$CA_ROOT/$caName/private/$caName.key"
 
@@ -152,9 +164,9 @@ function createCA()
 #
 function selfsignCA()
 {
-    local caName=$1
+    local caName="$1"
     local caConf="$CNF_ROOT/ca.$caName.conf"
-    local outFile=$CA_ROOT/$caName.crt
+    local outFile="$CA_ROOT/$caName.crt"
 
     if [[ $# -ne 1 ]]; then
         echo "Usage: ${FUNCNAME[0]} caName"
@@ -199,8 +211,8 @@ function selfsignCA()
 #
 function signCA()
 {
-    local caName=$1
-    local parentCaName=$2
+    local caName="$1"
+    local parentCaName="$2"
     local parentCaConf="$CNF_ROOT/ca.$parentCaName.conf"
     local outFile="$CA_ROOT/$caName.crt"
     local outChainFile="$CA_ROOT/$caName-chain.crt"
@@ -228,7 +240,7 @@ function signCA()
     setCertValidity
 
     # Create CA certificate
-    openssl ca -config "$parentCaConf" -in "$CA_ROOT/$caName.csr" -out "$outFile" -extensions signing_ca_ext -startdate "${START_DATE}000000Z}" -enddate "${END_DATE}235959Z"
+    openssl ca -config "$parentCaConf" -in "$CA_ROOT/$caName.csr" -out "$outFile" -extensions signing_ca_ext -startdate "${START_DATE}000000Z" -enddate "${END_DATE}235959Z"
 
     if [[ ! -s "$outFile" ]]; then
         # File is zero-sized!
@@ -254,7 +266,7 @@ function signCA()
 #
 function createCRL()
 {
-    local caName=$1
+    local caName="$1"
     local caConf="$CNF_ROOT/ca.$caName.conf"
     local outFile="$CRL_ROOT/$caName.crl"
 
@@ -286,22 +298,22 @@ function createCRL()
 # OpenSSL configuration file (of the CA to be used for signing) to
 # exist with the name conf/req.[caName].conf
 #
-# @param caName   The name of the CA to be used for signing later.
 # @param certName The name of the private key to be created. Note
 #                 that this specifies the file name. The name of the
 #                 owner as well as additional information will be
 #                 querried by OpenSSL on the command line.
+# @param caName   The name of the CA to be used for signing later.
 #
 function createClientKey()
 {
-    local caName=$1
+    local certName="$1"
+    local caName="$2"
     local caConf="$CNF_ROOT/req.$caName.conf"
-    local certName=$2
-    local outKey=$CERT_ROOT/$certName.pem
-    local outCsr=$CERT_ROOT/$certName.csr
+    local outKey="$CERT_ROOT/$certName.pem"
+    local outCsr="$CERT_ROOT/$certName.csr"
 
     if [[ $# -ne 2 ]]; then
-        echo "Usage: ${FUNCNAME[0]} caName certName"
+        echo "Usage: ${FUNCNAME[0]} certName caName"
         echo ""
         cahelp createClientKey
         echo "Run 'cahelp' for a documentation on all available functions"
@@ -321,29 +333,29 @@ function createClientKey()
     openssl req -new -config "$caConf" -key "$outKey" -out "$outCsr"
 }
 
+
 # Sign an existing client key using a specified CA. If any certificates
 # or certificate bundles already exist, the process will be aborted.
 #
 # This function expects an OpenSSL configuration file (of the CA to be
 # used for signing) to exist with the name conf/ca.[caName].conf
 #
-# @param caName   The name of the CA to be used for signing.
 # @param certName The name of the certificate signing request (without
 #                 extension)
 # @param name     The name of the owner to create the certificate for.
+# @param caName   The name of the CA to be used for signing.
 #
 function signCertificate()
 {
-    local caName=$1
+    local certName="$1"
+    local name="$2"
+    local caName="$3"
     local caConf="$CNF_ROOT/ca.$caName.conf"
-    local certName=$2
-    local name=$3
     local outFile="$CERT_ROOT/$certName.crt"
-    local outPKCSBundleFile="$CERT_ROOT/$certName.p12"
     local outPEMBundleFile="$CERT_ROOT/$certName.key+crt"
 
     if [[ $# -ne 3 ]]; then
-        echo "Usage: ${FUNCNAME[0]} caName certName name"
+        echo "Usage: ${FUNCNAME[0]} certName name caName"
         echo ""
         cahelp signCertificate
         echo "Run 'cahelp' for a documentation on all available functions"
@@ -354,21 +366,20 @@ function signCertificate()
     elif [[ -f "$outFile" ]]; then
         echo "Error: client certificate '$certName' already exists"
         return 1
-    elif [[ -f "$outPKCSBundleFile" ]]; then
-        echo "Error: client certificate PKCS#12 bundle '$certName' already exists"
-        return 1
     elif [[ -f "$outPEMBundleFile" ]]; then
         echo "Error: client certificate PEM bundle '$certName' already exists"
         return 1
-    elif [[ ! -f "$CA_ROOT/$certName.csr" ]]; then
+    elif [[ ! -f "$CERT_ROOT/$certName.csr" ]]; then
         echo "Error: certificate signing request for '$certName' does not exist"
         return 1
+    elif [[ ! -f "$CERT_ROOT/$certName.pem" ]]; then
+        echo "Warning: private key for '$certName' does not exist, bundle will not be created"
     fi
 
     setCertValidity
 
     # Create certificate
-    openssl ca -config "$caConf" -in "$CERT_ROOT/$certName.csr" -out "$outFile" -extensions ${caName}_ext -startdate "${START_DATE}000000Z}" -enddate "${END_DATE}235959Z"
+    openssl ca -config "$caConf" -in "$CERT_ROOT/$certName.csr" -out "$outFile" -extensions ${caName}_ext -startdate "${START_DATE}000000Z" -enddate "${END_DATE}235959Z"
 
     if [[ ! -s "$outFile" ]]; then
         # File is zero-sized!
@@ -377,12 +388,69 @@ function signCertificate()
         return 1
     fi
 
-    # Create PKCS#12 bundle
-    openssl pkcs12 -export -inkey "$CERT_ROOT/$certName.pem" -in "$outFile" -certfile "$CA_ROOT/$caName-chain.crt" -out "$outPKCSBundleFile" -name "$name"
+    if [[ -f "$CERT_ROOT/$certName.pem" ]]; then
+        # Create PEM bundle
+        cat "$CERT_ROOT/$certName.pem" "$outFile" > "$outPEMBundleFile"
+    fi
 
-    # Create PEM bundle
-    cat "$CERT_ROOT/$certName.pem" "$outFile" > "$outPEMBundleFile"
+    echo ""
+    echo "Creating a PKCS#12 bundle now. This will prompt for the password for the"
+    echo "private key. If that is not desired or the private key is not available,"
+    echo "you may exit the process by pressing Ctrl+C."
+    echo "To create the PKCS#12 bundle at a later time, run"
+    echo "  createPKCS12 $certName $caName"
+    echo ""
+
+    createPKCS12 $certName $caName
 }
+
+
+# Create a PKCS#12 certificate bundle containing the certificate and
+# private key (you will need to enter the password for the key).
+#
+# This function expects an OpenSSL configuration file (of the CA to be
+# used for signing) to exist with the name conf/ca.[caName].conf
+#
+# @param certName The name of the certificate
+# @param caName   The name of the certificate authority. This is needed
+#                 to attach the authority certificates to the PKCS#12
+#                 bundle. This is optional, but highly recommended.
+function createPKCS12()
+{
+    local certName="$1"
+    local caName="$2"
+    local certFile="$CERT_ROOT/$certName.crt"
+    local certKey="$CERT_ROOT/$certName.pem"
+    local caCertChain="$CA_ROOT/$caName-chain.crt"
+    local outPKCSBundleFile="$CERT_ROOT/$certName.p12"
+
+    if [[ $# -ne 1 && $# -ne 2 ]]; then
+        echo "Usage: ${FUNCNAME[0]} certName [caName]"
+        echo ""
+        cahelp createPKCS12
+        echo "Run 'cahelp' for a documentation on all available functions"
+        return 1
+    elif [[ -f "$outPKCSBundleFile" ]]; then
+        echo "Error: client certificate PKCS#12 bundle '$certName' already exists"
+        return 1
+    elif [[ ! -f "$certFile" ]]; then
+        echo "Error: certificate for '$certName' does not exist"
+        return 1
+    elif [[ ! -f "$certKey" ]]; then
+        echo "Error: private key for '$certName' does not exist"
+        return 1
+    elif [[ ! -f "$caCertChain" ]]; then
+        echo "Warning: certificate chain for the ca does not exist, skipping"
+    fi
+
+    # Create PKCS#12 bundle
+    if [[ -f "$caCertChain" ]]; then
+        openssl pkcs12 -export -inkey "$certKey" -in "$certFile" -certfile "$CA_ROOT/$caName-chain.crt" -out "$outPKCSBundleFile" -name "$name"
+    else
+        openssl pkcs12 -export -inkey "$certKey" -in "$certFile" -out "$outPKCSBundleFile" -name "$name"
+    fi
+}
+
 
 # Revoke an existing client certificate that was previously signed by a
 # CA.
@@ -397,9 +465,9 @@ function signCertificate()
 #
 function revokeCertificate()
 {
-    local caName=$1
+    local caName="$1"
     local caConf="$CNF_ROOT/ca.$caName.conf"
-    local certNumber=$2
+    local certNumber="$2"
 
     if [[ $# -ne 2 ]]; then
         echo "Usage: ${FUNCNAME[0]} caName certNumber"
